Added global service messages

+++ src/Systems/GlobalServiceMessages.cpp
@@ -48,6 +48,8 @@ GlobalServiceMessages::~GlobalServiceMessages()
 void GlobalServiceMessages::init(SharedObjects* baseLib)
 {
     _bl = baseLib;
+    _rpcDecoder = std::unique_ptr<BaseLib::Rpc::RpcDecoder>(new BaseLib::Rpc::RpcDecoder(baseLib, false, false));
+    _rpcEncoder = std::unique_ptr<BaseLib::Rpc::RpcEncoder>(new BaseLib::Rpc::RpcEncoder(baseLib, false, true));
 }
 
 void GlobalServiceMessages::load()
@@ -65,6 +67,7 @@ void GlobalServiceMessages::load()
             serviceMessage->timestamp = row.second.at(4)->intValue;
             serviceMessage->message = row.second.at(6)->textValue;
             serviceMessage->value = row.second.at(5)->intValue;
+            serviceMessage->data = _rpcDecoder->decodeResponse(*row.second.at(7)->binaryValue);
             _serviceMessages[row.second.at(1)->intValue].emplace(row.second.at(3)->intValue, std::move(serviceMessage));
         }
     }
@@ -82,7 +85,7 @@ void GlobalServiceMessages::load()
     }
 }
 
-void GlobalServiceMessages::set(int32_t familyId, int32_t messageId, int32_t timestamp, std::string message, int64_t value)
+void GlobalServiceMessages::set(int32_t familyId, int32_t messageId, int32_t timestamp, std::string message, PVariable data, int64_t value)
 {
     try
     {
@@ -92,11 +95,15 @@ void GlobalServiceMessages::set(int32_t familyId, int32_t messageId, int32_t tim
         serviceMessage->timestamp = timestamp;
         serviceMessage->message = message;
         serviceMessage->value = value;
+        serviceMessage->data = data;
 
         {
             std::lock_guard<std::mutex> serviceMessagesGuard(_serviceMessagesMutex);
             _serviceMessages[familyId][messageId] = std::move(serviceMessage);
         }
+
+        std::vector<char> dataBlob;
+        if(data) _rpcEncoder->encodeResponse(data, dataBlob);
 
         Database::DataRow data;
         data.push_back(std::make_shared<Database::DataColumn>(familyId));
@@ -105,7 +112,7 @@ void GlobalServiceMessages::set(int32_t familyId, int32_t messageId, int32_t tim
         data.push_back(std::make_shared<Database::DataColumn>(timestamp));
         data.push_back(std::make_shared<Database::DataColumn>(value));
         data.push_back(std::make_shared<Database::DataColumn>(message));
-        data.push_back(std::make_shared<Database::DataColumn>());
+        data.push_back(std::make_shared<Database::DataColumn>(dataBlob));
         _bl->db->saveGlobalServiceMessageAsynchronous(data);
     }
     catch(const std::exception& ex)
@@ -131,12 +138,16 @@ void GlobalServiceMessages::unset(int32_t familyId, int32_t messageId)
             auto familyIterator = _serviceMessages.find(familyId);
             if(familyIterator != _serviceMessages.end())
             {
-                familyIterator->second.erase(messageId);
-                if(familyIterator->second.empty()) _serviceMessages.erase(familyId);
+                auto messageIterator = familyIterator->second.find(messageId);
+                if(messageIterator != familyIterator->second.end())
+                {
+                    familyIterator->second.erase(messageIterator);
+                    if(familyIterator->second.empty()) _serviceMessages.erase(familyId);
+
+                    _bl->db->deleteGlobalServiceMessage(familyId, messageId);
+                }
             }
         }
-
-        _bl->db->deleteGlobalServiceMessage(familyId, messageId);
     }
     catch(const std::exception& ex)
     {
@@ -169,6 +180,7 @@ PVariable GlobalServiceMessages::get(PRpcClientInfo clientInfo)
                 element->structValue->emplace("TIMESTAMP", std::make_shared<Variable>(message.second->timestamp));
                 element->structValue->emplace("MESSAGE_ID", std::make_shared<Variable>(message.second->messageId));
                 element->structValue->emplace("MESSAGE", std::make_shared<Variable>(message.second->message));
+                element->structValue->emplace("DATA", message.second->data);
                 element->structValue->emplace("VALUE", std::make_shared<Variable>(message.second->value));
                 serviceMessages->arrayValue->push_back(element);
                 if(serviceMessages->arrayValue->size() == serviceMessages->arrayValue->capacity()) serviceMessages->arrayValue->reserve(serviceMessages->arrayValue->size() + 100);
